refactor: Rename wrapper to EditorWrapper
