+++ src/chainlit/frontend/src/components/organisms/playground/editor/EditorWrapper.tsx
