+++ src/chainlit/frontend/src/components/organisms/playground/editor/template/index.tsx
@@ -13,7 +13,7 @@ import {
 import { useState } from 'react';
 import { useIsFirstRender } from 'usehooks-ts';
 
-import EditorWrapper from 'components/organisms/playground/editor/wrapper';
+import EditorWrapper from 'components/organisms/playground/editor/EditorWrapper';
 
 import { IPrompt } from 'types/chat';
